rearranged some things & removed AvlTree.hpp
