+++ .gitignore
@@ -33,3 +33,6 @@
 
 # VS generated files
 .vs/
+
+# test files
+test/
