+++ includes.hpp
@@ -1,0 +1,5 @@
+// https://github.com/jarro2783/cxxopts/tree/2_1
+#include "cxxopts.hpp" // implicitly includes several STL containers
+#include "Map.hpp" // Avl tree implementation
+#include "MapSet.hpp"  // map of sets for search
+#include "PriorityQueue.hpp" // priority queue implementations
